Move search components to 'client/src/components/shared/search' package
