+++ client/src/pages/events.js
@@ -7,7 +7,7 @@ import Spinner from 'react-bootstrap/Spinner';
 import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
 import EventCard from '../components/event-search/event-card';
 import 'react-datepicker/dist/react-datepicker.css';
-import LocalSearch from '../components/local-search';
+import LocalSearch from '../components/shared/search/local-search';
 import DropdownGroup from '../components/event-search/dropdown-group';
 import Api from '../services/api';
